fix bug with missing target for OTM dataflows
